+++ server/routes/eatcoin-api/categories.js
@@ -1,0 +1,9 @@
+var express = require('express');
+var router = express.Router();
+
+/* GET home page. */
+router.get('/', function(req, res, next) {
+  res.send('categories-api')  // render: 응답할 페이지 렌더링
+});
+
+module.exports = router;
